requirements.txt
openai==1.51.0
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
streamlit==1.39.0
streamlit-mermaid==0.2.0